pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "english-coach"
version = "0.9.3"
description = "AI-powered English tutor for STEM students — TOEFL/GRE/IELTS/CET"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pywebview>=5.0.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
full = [
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "language-tool-python>=2.8.0",
    "edge-tts>=6.1.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
english-coach = "cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["core", "ai", "modes", "cli"]
